bumps schema engine to v9, adds migrations
