+++ tests/common/storages/test_schema_storage.py
@@ -4,9 +4,9 @@ import pytest
 import yaml
 from dlt.common import json
 
+from dlt.common.normalizers import explicit_normalizers
 from dlt.common.schema.schema import Schema
 from dlt.common.schema.typing import TStoredSchema
-from dlt.common.schema.utils import explicit_normalizers
 from dlt.common.storages.exceptions import (
     InStorageSchemaModified,
     SchemaNotFoundError,
@@ -24,7 +24,7 @@ from tests.common.utils import (
     load_yml_case,
     yml_case_path,
     COMMON_TEST_CASES_PATH,
-    IMPORTED_VERSION_HASH_ETH_V8,
+    IMPORTED_VERSION_HASH_ETH_V9,
 )
 
 
@@ -227,10 +227,10 @@ def test_save_store_schema_over_import(ie_storage: SchemaStorage) -> None:
     ie_storage.save_schema(schema)
     assert schema.version_hash == schema_hash
     # we linked schema to import schema
-    assert schema._imported_version_hash == IMPORTED_VERSION_HASH_ETH_V8
+    assert schema._imported_version_hash == IMPORTED_VERSION_HASH_ETH_V9
     # load schema and make sure our new schema is here
     schema = ie_storage.load_schema("ethereum")
-    assert schema._imported_version_hash == IMPORTED_VERSION_HASH_ETH_V8
+    assert schema._imported_version_hash == IMPORTED_VERSION_HASH_ETH_V9
     assert schema._stored_version_hash == schema_hash
     assert schema.version_hash == schema_hash
     assert schema.previous_hashes == []
@@ -247,7 +247,7 @@ def test_save_store_schema_over_import_sync(synced_storage: SchemaStorage) -> No
     schema = Schema("ethereum")
     schema_hash = schema.version_hash
     synced_storage.save_schema(schema)
-    assert schema._imported_version_hash == IMPORTED_VERSION_HASH_ETH_V8
+    assert schema._imported_version_hash == IMPORTED_VERSION_HASH_ETH_V9
     # import schema is overwritten
     fs = FileStorage(synced_storage.config.import_schema_path)
     exported_name = synced_storage._file_name_in_store("ethereum", "yaml")
@@ -327,13 +327,13 @@ def prepare_import_folder(storage: SchemaStorage) -> None:
 
 def assert_schema_imported(synced_storage: SchemaStorage, storage: SchemaStorage) -> Schema:
     prepare_import_folder(synced_storage)
-    eth_V8: TStoredSchema = load_yml_case("schemas/eth/ethereum_schema_v8")
+    eth_V9: TStoredSchema = load_yml_case("schemas/eth/ethereum_schema_v9")
     schema = synced_storage.load_schema("ethereum")
     # is linked to imported schema
-    schema._imported_version_hash = eth_V8["version_hash"]
+    schema._imported_version_hash = eth_V9["version_hash"]
     # also was saved in storage
     assert synced_storage.has_schema("ethereum")
     # and has link to imported schema s well (load without import)
     schema = storage.load_schema("ethereum")
-    assert schema._imported_version_hash == eth_V8["version_hash"]
+    assert schema._imported_version_hash == eth_V9["version_hash"]
     return schema
